requirements.txt
PySimpleGUIQt~=0.35.0
PyYAML~=5.3.1
ahk~=0.11.1